Move file system export back
