+++ src/Package/Impl/ProjectSystem/Exports.cs
@@ -2,8 +2,9 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System.ComponentModel.Composition;
+using Microsoft.Common.Core.IO;
 
-namespace Microsoft.Common.Core.IO {
+namespace Microsoft.VisualStudio.R.Package.ProjectSystem {
     internal sealed class Export {
 
         [Export(typeof(IFileSystem))]
